Mesh file impl in mesh mod
